修改一些接口文件和request封装
